+++ tests/controller_base_test.py
@@ -1,4 +1,5 @@
 import pytest
+from hycon.controllers import WindFarmPowerTrackingController
 from hycon.controllers.controller_base import ControllerBase
 
 
@@ -45,3 +46,10 @@ def test_inherited_methods(test_interface_standin):
         _ = InheritanceTestClassBad(test_interface_standin)
 
     _ = InheritanceTestClassGood(test_interface_standin)
+
+
+def test_inherited_instantiation(test_interface_standin):
+    """
+    Check that a subclass of InterfaceBase can be instantiated.
+    """
+    _ = WindFarmPowerTrackingController(interface=test_interface_standin, cname="test_cname")
