+++ tests/hybrid_controllers_test.py
@@ -1,173 +1,15 @@
 import copy
 
 import numpy as np
-import pandas as pd
 import pytest
 from hycon.controllers import (
     BatteryPassthroughController,
     HybridSupervisoryControllerGeneric,
     HydrogenPlantController,
-    LookupBasedWakeSteeringController,
-    PriceCurtailingController,
     SolarPassthroughController,
     WindFarmPowerDistributingController,
     WindFarmPowerTrackingController,
 )
-from hycon.controllers.wind_farm_power_tracking_controller import POWER_SETPOINT_DEFAULT
-
-
-def test_controller_instantiation(test_interface_standin):
-    """
-    Tests whether all controllers can be imported correctly and that they
-    each implement the required methods specified by ControllerBase.
-    """
-    _ = LookupBasedWakeSteeringController(interface=test_interface_standin, cname="wind_farm")
-    _ = WindFarmPowerDistributingController(interface=test_interface_standin, cname="wind_farm")
-    _ = WindFarmPowerTrackingController(interface=test_interface_standin, cname="wind_farm")
-    _ = SolarPassthroughController(interface=test_interface_standin, cname="solar_farm")
-
-
-def test_LookupBasedWakeSteeringController(test_hercules_v1_dict, test_interface_hercules_ad):
-    # No lookup table passed; simply passes through wind direction to yaw angles
-    test_controller = LookupBasedWakeSteeringController(
-        interface=test_interface_hercules_ad, cname="wind_farm"
-    )
-
-    # Check that the controller can be stepped
-    test_hercules_v1_dict["time"] = 20
-    test_dict_out = test_controller.step(input_dict=test_hercules_v1_dict)
-    test_angles = np.array(
-        test_dict_out["hercules_comms"]["amr_wind"]["test_farm"]["turbine_yaw_angles"]
-    )
-    wind_directions = np.array(
-        test_hercules_v1_dict["hercules_comms"]["amr_wind"]["test_farm"]["turbine_wind_directions"]
-    )
-    assert np.allclose(test_angles, wind_directions)
-
-    # Lookup table that specified 20 degree offset for T000, 10 degree offset for T001 for all
-    # wind directions
-    test_offsets = np.array([20.0, 10.0])
-    df_opt_test = pd.DataFrame(
-        data={
-            "wind_direction": [220.0, 220.0, 320.0, 320.0],
-            "wind_speed": [0.0, 20.0, 0.0, 20.0],
-            "yaw_angles_opt": [test_offsets] * 4,
-            "turbulence_intensity": [0.06] * 4,
-        }
-    )
-    test_controller = LookupBasedWakeSteeringController(
-        interface=test_interface_hercules_ad,
-        cname="wind_farm",
-        controller_parameters={"df_yaw": df_opt_test},
-    )
-
-    test_hercules_v1_dict["time"] = 20
-    test_dict_out = test_controller.step(input_dict=test_hercules_v1_dict)
-    test_angles = np.array(
-        test_dict_out["hercules_comms"]["amr_wind"]["test_farm"]["turbine_yaw_angles"]
-    )
-    wind_directions = np.array(
-        test_hercules_v1_dict["hercules_comms"]["amr_wind"]["test_farm"]["turbine_wind_directions"]
-    )
-    assert np.allclose(test_angles, wind_directions - test_offsets)
-
-
-def test_WindFarmPowerDistributingController(test_hercules_v1_dict, test_interface_hercules_ad):
-    test_controller = WindFarmPowerDistributingController(
-        interface=test_interface_hercules_ad, cname="wind_farm"
-    )
-
-    # Default behavior when no power reference is given
-    test_hercules_v1_dict["time"] = 20
-    test_hercules_v1_dict["external_signals"] = {}
-    test_dict_out = test_controller.step(input_dict=test_hercules_v1_dict)
-    test_power_setpoints = np.array(
-        test_dict_out["hercules_comms"]["amr_wind"]["test_farm"]["turbine_power_setpoints"]
-    )
-    assert np.allclose(
-        test_power_setpoints,
-        POWER_SETPOINT_DEFAULT / 2,
-    )
-
-    # Test with power reference
-    test_hercules_v1_dict["external_signals"]["wind_power_reference"] = 1000
-    test_dict_out = test_controller.step(input_dict=test_hercules_v1_dict)
-    test_power_setpoints = np.array(
-        test_dict_out["hercules_comms"]["amr_wind"]["test_farm"]["turbine_power_setpoints"]
-    )
-    assert np.allclose(test_power_setpoints, 500)
-
-    # Test that ramp rate limits are applied
-    test_controller = WindFarmPowerDistributingController(
-        interface=test_interface_hercules_ad,
-        cname="wind_farm",
-        controller_parameters={"ramp_rate_limit": 200},
-    )
-    test_hercules_v1_dict["external_signals"]["wind_power_reference"] = 1000
-    test_controller.step(input_dict=test_hercules_v1_dict)  # To initialize previous power setpoints
-    test_hercules_v1_dict["external_signals"]["wind_power_reference"] = 500
-    test_dict_out = test_controller.step(input_dict=test_hercules_v1_dict)
-    test_power_setpoints = np.array(
-        test_dict_out["hercules_comms"]["amr_wind"]["test_farm"]["turbine_power_setpoints"]
-    )
-    assert np.allclose(test_power_setpoints, (1000 - 200) / 2)
-
-    test_hercules_v1_dict["external_signals"]["wind_power_reference"] = 2000
-    test_dict_out = test_controller.step(input_dict=test_hercules_v1_dict)
-    test_power_setpoints = np.array(
-        test_dict_out["hercules_comms"]["amr_wind"]["test_farm"]["turbine_power_setpoints"]
-    )
-    assert np.allclose(test_power_setpoints, 1000 / 2)
-
-
-def test_WindFarmPowerTrackingController(test_hercules_v1_dict, test_interface_hercules_ad):
-    test_controller = WindFarmPowerTrackingController(
-        interface=test_interface_hercules_ad, cname="wind_farm"
-    )
-
-    # Test no change to power setpoints if producing desired power
-    test_hercules_v1_dict["external_signals"]["wind_power_reference"] = 1000
-    test_hercules_v1_dict["hercules_comms"]["amr_wind"]["test_farm"]["turbine_powers"] = [500, 500]
-    test_dict_out = test_controller.step(input_dict=test_hercules_v1_dict)
-    test_power_setpoints = np.array(
-        test_dict_out["hercules_comms"]["amr_wind"]["test_farm"]["turbine_power_setpoints"]
-    )
-    assert np.allclose(test_power_setpoints, 500)
-
-    # Test if power exceeds farm reference, power setpoints are reduced
-    test_hercules_v1_dict["hercules_comms"]["amr_wind"]["test_farm"]["turbine_powers"] = [600, 600]
-    test_dict_out = test_controller.step(input_dict=test_hercules_v1_dict)
-    test_power_setpoints = np.array(
-        test_dict_out["hercules_comms"]["amr_wind"]["test_farm"]["turbine_power_setpoints"]
-    )
-    assert (
-        test_power_setpoints
-        <= test_hercules_v1_dict["hercules_comms"]["amr_wind"]["test_farm"]["turbine_powers"]
-    ).all()
-
-    # Test if power is less than farm reference, power setpoints are increased
-    test_hercules_v1_dict["hercules_comms"]["amr_wind"]["test_farm"]["turbine_powers"] = [550, 400]
-    test_dict_out = test_controller.step(input_dict=test_hercules_v1_dict)
-    test_power_setpoints = np.array(
-        test_dict_out["hercules_comms"]["amr_wind"]["test_farm"]["turbine_power_setpoints"]
-    )
-    assert (
-        test_power_setpoints
-        >= test_hercules_v1_dict["hercules_comms"]["amr_wind"]["test_farm"]["turbine_powers"]
-    ).all()
-
-    # Test that more aggressive control leads to faster response
-    test_controller = WindFarmPowerTrackingController(
-        interface=test_interface_hercules_ad,
-        cname="wind_farm",
-        controller_parameters={"proportional_gain": 2},
-    )
-    test_hercules_v1_dict["hercules_comms"]["amr_wind"]["test_farm"]["turbine_powers"] = [600, 600]
-    test_dict_out = test_controller.step(input_dict=test_hercules_v1_dict)
-    test_power_setpoints_a = np.array(
-        test_dict_out["hercules_comms"]["amr_wind"]["test_farm"]["turbine_power_setpoints"]
-    )
-    assert (test_power_setpoints_a < test_power_setpoints).all()
 
 
 def test_HybridSupervisoryControllerGeneric_reference_tracking(
@@ -449,74 +291,3 @@ def test_HydrogenPlantController(test_hercules_dict, test_interface_hercules):
             interface=test_interface_hercules,
             controller_parameters=external_controller_parameters,
         )
-
-
-def test_PriceCurtailingController(test_hercules_dict, test_interface_hercules):
-    """
-    Tests that the PriceCurtailingController outputs a reasonable signal
-    """
-    # Consider a solar farm only
-    test_interface_hercules.component_names = ["solar_farm"]
-    test_controller = PriceCurtailingController(
-        interface=test_interface_hercules,
-        cname="solar_farm",
-        controller_parameters={
-            "curtailment_price": 50,
-            "power_tracking_controller": SolarPassthroughController(
-                test_interface_hercules, "solar_farm"
-            ),
-        },
-    )
-
-    # Test with price above curtailment threshold
-    power_setpoint_ref = 1000
-    test_hercules_dict["external_signals"]["lmp_rt"] = 100
-    test_hercules_dict["solar_farm"]["power_reference"] = power_setpoint_ref
-    out_dict = test_controller.step(test_hercules_dict)
-    power_setpoint_test = np.array(out_dict["solar_farm"]["power_setpoint"])
-    assert np.isclose(power_setpoint_test, power_setpoint_ref)
-
-    # Test with price below curtailment threshold
-    test_hercules_dict["external_signals"]["lmp_rt"] = 25
-    out_dict = test_controller.step(test_hercules_dict)
-    power_setpoint_test = np.array(out_dict["solar_farm"]["power_setpoint"])
-    assert np.isclose(power_setpoint_test, 0)
-
-    # Test again with negative threshold
-    test_controller.set_controller_parameters(
-        curtailment_price=-10,
-        power_tracking_controller=SolarPassthroughController(test_interface_hercules, "solar_farm"),
-    )
-    test_hercules_dict["external_signals"]["lmp_rt"] = -5
-    test_hercules_dict["solar_farm"]["power_reference"] = power_setpoint_ref
-    out_dict = test_controller.step(test_hercules_dict)
-    power_setpoint_test = np.array(out_dict["solar_farm"]["power_setpoint"])
-    assert np.isclose(power_setpoint_test, power_setpoint_ref)
-
-    test_hercules_dict["external_signals"]["lmp_rt"] = -15
-    out_dict = test_controller.step(test_hercules_dict)
-    power_setpoint_test = np.array(out_dict["solar_farm"]["power_setpoint"])
-    assert np.isclose(power_setpoint_test, 0)
-
-    # Test with wind farm
-    test_interface_hercules.component_names = ["wind_farm"]
-    test_controller = PriceCurtailingController(
-        interface=test_interface_hercules,
-        cname="wind_farm",
-        controller_parameters={
-            "curtailment_price": 50,
-            "power_tracking_controller": WindFarmPowerDistributingController(
-                test_interface_hercules, "wind_farm"
-            ),
-        },
-    )
-    test_hercules_dict["external_signals"]["lmp_rt"] = 100
-    test_hercules_dict["wind_farm"]["power_reference"] = power_setpoint_ref
-    out_dict = test_controller.step(test_hercules_dict)
-    power_setpoint_test = sum(out_dict["wind_farm"]["turbine_power_setpoints"])
-    assert np.isclose(power_setpoint_test, power_setpoint_ref)
-
-    test_hercules_dict["external_signals"]["lmp_rt"] = 25
-    out_dict = test_controller.step(test_hercules_dict)
-    power_setpoint_test = sum(out_dict["wind_farm"]["turbine_power_setpoints"])
-    assert np.isclose(power_setpoint_test, 0)
